backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
pydantic==2.10.6
PyJWT==2.10.1
python-dateutil==2.9.0.post0
